[add] mechanics to react on message reply with a Signal
The logic is the following:
* we send a message;
* if the user replies to it (with reply command or just sending their
message right after ours), the bus issues a signal;
* we then can handle this signal.

At last we can implement notes editing etc.

The signal must be all set up before the message is sent. This is
better than to try to guess what parameters should it get.

Maybe later we'll allow callables which produce signals based on
context, but for now things are already complex enough.