[fix] auto guess language and disable it via config
The guessing is fine when alphabets differ, but when the languages
share something (e.g. English vs German), the quality is terrible: the
classification result may be ok, but the confidence rarely gets above
0.9. Since we don't provide tools to fix the classification error and
we even don't have a command to explicitly translate *to* (only !! to
translate *from*), we shouldn't enable this for now.