[refactor/fix] maturity levels
Previous maturity was unstable because it used the cards scheduled
date. When this date became close, the card turned from mature to
young.

Maturity is an internal property of a card, not a function of how
close the rehearse date is.

So now the card is called mature if its stability (the number of days
it takes for the retrievability to drop from 100% to 90%) is more than
7 days (the default value, configurable).

This also fixes the annoying bug when not all notes were shown in the
notes list (due to faulty old conditions which were not exhaustive
altogether).

Also the performance tests on `get_notes` function were added.